requirements.txt
pygame==1.9.3
PyTMX==3.21.5
six==1.11.0
PyYAML==3.12
pyscroll==2.17.8
-e git+https://github.com/bitcraft/animation.git#egg=pygame-animation